Updating things to compile

+++ Goblin Camp/src/Game.cpp
@@ -840,7 +840,7 @@ void Game::GenerateMap() {
 	int lo_offset = 0;
 	int hi_offset = 0;
 	for (int x = -25; x < 25; ++x) {
-		int range = int(std::sqrt(25*25 - x*x));
+		int range = int(std::sqrt((double)(25*25 - x*x)));
 		lo_offset = std::min(std::max(rand() % 3 - 1 + lo_offset, -5), 5);
 		hi_offset = std::min(std::max(rand() % 3 - 1 + hi_offset, -5), 5);
 		for (int y = -range-lo_offset; y < range+hi_offset; ++y) {
